requirements.txt
pandas
numpy
psycopg2-binary
streamlit
plotly
python-dotenv
